requirements.txt
fastapi
uvicorn
pandas
spacy
transformers==4.43.3
torch
numpy
azure-storage-blob
python-dotenv